Finalized Public Release
Optimized .md files with proper link references and such!
Finalized a Importable npm module in typescript as well published under the npm name of pheonix-box
while still having the npm cli available to install globally using the npm module pheonix-box-cli
along with the release of the Binary CLI as well! All in a single Github Repository for Release!